requirements.txt
# --------------------------------------------------
# CORE BACKEND
# --------------------------------------------------
supabase==2.3.0
gotrue==2.9.1
postgrest==0.13.2
python-dotenv==1.0.1
fastapi==0.115.12
uvicorn==0.34.0

# --------------------------------------------------
# HTTP STACK (CRITICAL FIX)
# --------------------------------------------------
httpx==0.24.1
httpcore==0.17.2
h11==0.14.0

# --------------------------------------------------
# DATA & NUMERICAL
# --------------------------------------------------
numpy==1.26.4
pandas==2.2.2

# --------------------------------------------------
# MACHINE LEARNING
# --------------------------------------------------
xgboost==2.0.3
scikit-learn==1.4.2

# --------------------------------------------------
# TRAFFIC DATA SOURCES
# --------------------------------------------------
WazeRouteCalculator==0.15

# --------------------------------------------------
# HTTP REQUESTS
# --------------------------------------------------
requests==2.31.0

# --------------------------------------------------
# SCHEDULER
# --------------------------------------------------
schedule==1.2.1
